% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdate00.m.R
\name{bdate00.m}
\alias{bdate00.m}
\title{Estimation of Multiple break dates}
\usage{
bdate00.m(y, x, n.size = 1, m, trim.size, vec.long)
}
\arguments{
\item{y}{A vector of dependent variables (NT x 1)}

\item{x}{A matrix of regressors (NT x p)}

\item{n.size}{The size of cross sections (N)}

\item{m}{The maximum number of breaks}

\item{trim.size}{The trimming parameter (to exclude the boundary of the sample)}

\item{vec.long}{SOMETING}
}
\value{
the estimated break dates given 1-m breaks
}
\description{
Break dates estimation from 1 to m breaks in quantile regression
}
\details{
This procedure estimates the model allowing for multiple breaks.
When m = 1, the procedure, "partition", is used.
When m > 1, this function applies the dynamic programming
algorithm to look for the optimal breaks.
}
\references{
Oka, T., and Qu, Z. (2011).
"Estimating structural changes in regression quantiles."
Journal of Econometrics, 162(2), 248-267.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
