% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.date.m.R
\name{ci.date.m}
\alias{ci.date.m}
\title{Confidence intervals for estimated break dates}
\usage{
ci.date.m(y, x, vec.tau, vec.date, n.size = 1, v.b = 2)
}
\arguments{
\item{y}{A vector of dependent variable (NT x 1)}

\item{x}{A matrix of regressors (NT x p)}

\item{vec.tau}{A set of probablities for conditioanl quantiles of interest}

\item{vec.date}{Estimated break dates}

\item{n.size}{The size of cross sections (N)}

\item{v.b}{v.b =1 for the 90\% confidence interval or =2 for the 95\% confidence interval \link{v.b=2 by default}}
}
\value{
A matrix in which the 1st column includes break dates; 2nd and 3rd columns include confidence intervals
}
\description{
Confidence intervals for break dates
}
\details{
This function constructs confidence intervals for estimated break dates
based on multiple regression quantiles.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
