% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.date.msub.R
\name{ci.date.m.sub}
\alias{ci.date.m.sub}
\title{Sub-rutine for ci.date.m}
\usage{
ci.date.m.sub(y.L, x.L, y.R, x.R, v.date, mat.size, vec.tau, n.size = 1, v.b)
}
\arguments{
\item{y.L}{observations (y) on the right-side of the break date}

\item{x.L}{observations (x) on the right-side of the break date}

\item{y.R}{observations (y) on the left-side of the break date}

\item{x.R}{observations (x) on the left-side of the break date}

\item{v.date}{break date}

\item{mat.size}{EXPLAIN}

\item{vec.tau}{Quantiles of interest}

\item{n.size}{The size of cross sections (N) \link{N=1 by default: time series}}

\item{v.b}{=1 for the 90\% confidence interval or =2 for the 95\% confidence interval}
}
\value{
A vector whose 1st column includes a break date, and 2nd and 3rd columns include confidence intervals
}
\description{
Subrutine for ci.date.m
}
\details{
This is a sub-rutine for ci.date.m.
Related routines include "Resample.split" and "moment".
}
\author{
Tatsushi Oka and Zhongjun Qu
}
