% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical.DQtest_specific.R
\name{critical.DQtest_specific}
\alias{critical.DQtest_specific}
\title{Sub-rutine for ci.date.m}
\usage{
critical.DQtest_specific(x, m.max, vec.tau)
}
\arguments{
\item{x}{the matrix of regressors (not including an intercept)}

\item{vec.tau}{Quantiles of interest}

\item{m}{the maximum number of breaks}
}
\value{
Output saved in your working directory.
}
\description{
critical.DQtes_specific
}
\details{
This procedure obtains critical values via simulation.
This procedure saves a text file in your working directory.
The name of the file is "table.cv.DQ_t1_t2_p.txt",
where
t1 = min(vec.tau), the minimum value of the quantiles;
t2 = max(vec.tau), the maximum value of the quantiles;
p  = the number of regressors (including an intercept).
}
\author{
Tatsushi Oka and Zhongjun Qu
}
