% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq.res.surface.R
\name{dq.res.surface.r}
\alias{dq.res.surface.r}
\title{Critical values using a response surface analysis}
\usage{
dq.res.surface.r(p, l, q.L, q.R)
}
\arguments{
\item{p}{The number of parameters in the model}

\item{l}{The number of changes under H0, i.e. l+1 under H1}

\item{q.L}{The left end of a range of quantiles}

\item{q.R}{The right end of a range of quantiles}
}
\value{
Critical value
}
\description{
Critical values using a response surface analysis
}
\details{
This procedure returns critical values using a response surface analysis.
Please be advised that this procedure only covers
the case where the trimming is symmetric, i.e., \link{tau,1-tau}.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
