% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{The Data Set for the Real GDP Growth analysis}
\format{
A data object with four variables
}
\usage{
data(gdp)
}
\description{
The data set contains the quaterly real GDP growth in the US.
The sample period is from 1947 Q4 to to 2009 Q2.
}
\details{
\itemize{
\item yq:     Year and quater ("Year Qauter" format: e.g. "1947 Q2")
\item gdp:    Real GDP growth rate (annualized by multiplying 4)
\item lag1:   One quater lagged "gdp"
\item lag2:   Two quaters lagged "gdp"
}
}
\examples{
data(gdp)
names(gdp)
summary(gdp)

}
\references{
Oka, T., and Qu, Z. (2011).
"Estimating structural changes in regression quantiles."
\emph{Journal of Econometrics}, 162(2), 248-267.
}
\keyword{datasets}
