% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.long.R
\name{gen.long}
\alias{gen.long}
\title{gen long}
\usage{
gen.long(y, x, vec.tau, n.size = 1, trim.size)
}
\arguments{
\item{y}{A dependent (NT x 1)}

\item{x}{A matrix of regressors (NT x p)}

\item{vec.tau}{Quantiles of interest}

\item{n.size}{The size of cross sections; default set to 1 #update 11/06/2013}

\item{trim.size}{Triming size}
}
\value{
Break date estimates
}
\description{
Subroutine for dynamic programing algorithm
}
\details{
This function obtains the objective function values given permissible regimes.
based on multiple quantiles.
Related subrutines are rq.est.full, sample.split, moment.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
