% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.mat.rho.R
\name{gen.mat.rho}
\alias{gen.mat.rho}
\title{Confidence intervals for break dates}
\usage{
gen.mat.rho(y, x, vec.tau, n.size = 1, trim.size, start, last)
}
\arguments{
\item{y}{A vector of dependent variables (NT x 1)}

\item{x}{A matrix of regressors (NT x p)}

\item{vec.tau}{Quantiles of interest}

\item{n.size}{The size of cross sections (n.size=1 by default)}

\item{trim.size}{The triming size}

\item{last}{SOMETHING}

\item{vec.date}{Estimated break dates}
}
\value{
A matrix in which the 1st column includes break dates; 2nd and 3rd columns include confidence intervals
}
\description{
Confidence intervals for break dates
}
\details{
This procedure obtains the values of the objective function
given permissible partitions.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
