% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.cv.dq.R
\name{get.cv.dq}
\alias{get.cv.dq}
\title{Ciritical Values for DQ test}
\usage{
get.cv.dq(n.break, p.size, q.L, q.R, d.Sym = FALSE)
}
\arguments{
\item{n.break}{the number of breaks under the null hypothesis.}

\item{p.size}{the number of regressors}

\item{q.L}{The lower end of the range of quantiles}

\item{q.R}{The upper end of the range of quantiles}

\item{d.Sym}{}
}
\value{
a 3x1 vector of critical values at 10\%, 5\%, 1\% significance level
}
\description{
Ciritical Values for the DQ test
}
\details{
This function calculates ciritical values for sequential DQ test for
the null hypothesis of L breaks and the alternatative hypothesis of L+1 breaks.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
