% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.cv.sq.R
\name{get.cv.sq}
\alias{get.cv.sq}
\title{Critical values for the SQ test}
\usage{
get.cv.sq(n.break, p.size)
}
\arguments{
\item{n.break}{the number of break under the null hypothesis}

\item{p.size}{the number of regressors}
}
\value{
a 3x1 vector of critical values at at 10\%, 5\%, 1\% significance level
}
\description{
Critical values for the SQ test
}
\details{
Critical values for the SQ test, SQ(l|l+1).
}
\author{
Tatsushi Oka and Zhongjun Qu
}
