% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Optimal Break Partition}
\usage{
partition(vec.long, start, b1, b2, last, t.size)
}
\arguments{
\item{vec.long}{a long vector contains the values of the objective function}

\item{start}{the beginning of the segment considered}

\item{b1}{the first possible break date}

\item{b2}{the last  possible break date}

\item{last}{the end of segment considered}

\item{t.size}{The size of time series}
}
\value{
loc.min the optimal break point

rho.min the value of objective function at the optimal break point
}
\description{
Optimal break partition
}
\details{
procedure to obtain an optimal one break partition
a segment, \link{start,end}.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
