% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR.break-package.R
\docType{package}
\name{QR.break-package}
\alias{QR.break}
\alias{QR.break-package}
\title{QR.break: Structural Breaks in Quantile Regression}
\description{
Estimation and inference methods for quantile regression with multiple structural breaks. [Complete this with a full description of your package functionality.]
}
\author{
\strong{Maintainer}: Tatsushi Oka \email{oka.econ@gmail.com}

Authors:
\itemize{
  \item Zhongjun Qu \email{qu@bu.edu}
}

}
\keyword{internal}
