% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rq.est.full.R
\name{rq.est.full}
\alias{rq.est.full}
\title{Estimation of a linear quantile regression with specific breask dates.}
\usage{
rq.est.full(y, x, v.tau, vec.date, n.size = 1)
}
\arguments{
\item{y}{A vector of dependent ($NT \times 1$)}

\item{x}{A matrix of regressors (NT x p)}

\item{v.tau}{the (single) quantile of interest}

\item{vec.date}{A vector of break dates, sepcified by users.}

\item{n.size}{The size of cross sections (N)}
}
\value{
object from quantile regression estimates, rq(), with structural breaks.
}
\description{
Estimation of a linear quantile regression with structual breaks
}
\details{
This procedure estimates a linear quantile regression,
given a set of sepcified break dates.
}
\references{
Koenker, R., and Bassett Jr, G. (1978).
"Regression quantiles". Econometrica, 46(1), 33-50.

Oka, T., and Qu, Z. (2011).
"Estimating structural changes in regression quantiles."
Journal of Econometrics, 162(2), 248-267.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
