% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.split.R
\name{sample.split}
\alias{sample.split}
\title{Sample splitting subroutine}
\usage{
sample.split(y, x, v.date, n.size = 1)
}
\arguments{
\item{y}{A vector of dependent variables (NT x 1)}

\item{x}{A matrix of regressors (NT x p)}

\item{v.date}{a break date}

\item{n.size}{The size of cross sections}
}
\value{
y1 a vector of dependent variables in the first sample

x1 a matrix of regressors in the first sample

y2 a vector of dependent variables in the second sample

x2 a matrix of regressors in the second sample
}
\description{
Subroutine: sample splitting
}
\details{
This procedure splits the sample into two subsamples.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
