% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.dq.R
\name{seq.dq}
\alias{seq.dq}
\title{The prcedure based on the DQ test to determine the number of breaks}
\usage{
\method{seq}{dq}(y, x, vec.tau, q.L, q.R, n.size = 1, m.max, trim.size, mat.date, d.Sym)
}
\arguments{
\item{y}{A vector of dependent variables (NT x 1)}

\item{x}{A matrix of regressors (NT x p)}

\item{vec.tau}{Quantiles of interest}

\item{q.L}{the quantile range (the left end) for the DQ test}

\item{q.R}{the quantile range (the right end) for the DQ test}

\item{n.size}{The size of cross sections (N)}

\item{m.max}{the maximum number of breaks allowed}

\item{trim.size}{a trimming value}

\item{mat.date}{a matrix of break dates}
}
\value{
test DQ test

cv critical values for DQ test (10\%, 5\% and 1\% signficance levels)

date estimated break dates

nbreak the number of break dates
}
\description{
This procedure determines the number of breaks using DQ(l|l+1) test
}
\details{
This procedure determines the number of breaks by applying the DQ test
sequentially.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
