% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.sq.R
\name{seq.sq}
\alias{seq.sq}
\title{The procedure based on the SQ test to determine the number of breaks}
\usage{
\method{seq}{sq}(y, x, v.tau, n.size = 1, m.max, trim.size, mat.date)
}
\arguments{
\item{y}{A vector of dependent (NT x 1)}

\item{x}{A matrix of regressors (NT x p)}

\item{v.tau}{A quantile of interest}

\item{n.size}{The size of cross sections (N)}

\item{m.max}{The maximum number of breaks allowed}

\item{trim.size}{The trimming size}

\item{mat.data}{a matrix of break dates}
}
\value{
test   \link{1xm.max}: DQ test

cv     \link{3xm.max}: critical values for DQ test; 1st, 2nd & 3rd row = 10, 5, 1\% size

date   \link{3xm.max}: break dates;
1st, 2nd & 3rd row = 10, 5, 1\% size
}
\description{
To determine the number of breaks using the SQ(l|l+1) test
}
\details{
This procedure determines the number of breaks by applying the SQ test
sequentially.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
