% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.prob.DQ.R
\name{vec.prob.DQ}
\alias{vec.prob.DQ}
\title{Subroutine to generete critical values of the DQ test}
\usage{
vec.prob.DQ(n.grid, n.dim, n.sim, vec.tau)
}
\arguments{
\item{n.grid}{the number of grid points}

\item{n.dim}{the number of repetition}

\item{n.sim}{the number of simulation}
}
\value{
Prob
}
\description{
This is a subroutine to generate critical values for the DQ test.
}
\details{
This is a subroutine to generate critical values for the DQ test.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
