% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq.test.0vs1.R
\name{dq.test.0vs1}
\alias{dq.test.0vs1}
\title{DQ test (test for single break over a range of quantiles)}
\usage{
dq.test.0vs1(y, x, q.L, q.R, n.size = 1)
}
\arguments{
\item{y}{A vector of dependent (NT x 1)}

\item{x}{A matrix of regressors (NT x (p-1)), excluding the constant term}

\item{q.L}{The lower end of the range of quantiles}

\item{q.R}{The upper end of the range of quantiles}

\item{n.size}{The size of cross sections (N)}

\item{vec.tau}{Quantiles of interest}

\item{vec.date}{Estimated break dates}
}
\value{
The value of the DQ test
}
\description{
Test for a single break over a range of quantiles
}
\details{
This procedure computes test statistics for the existence of a single break
over an interval of quantiles. The null hypothesis is no break
in linear quantile regressions given the interval of quantiles,
while the alternative hypothesis is the existence of breaks.
}
\references{
Koenker, R., and Bassett Jr, G. (1978).
"Regression quantiles". Econometrica, 46(1), 33-50.

Qu, Z. (2008).
"Testing for structural change in regression quantiles."
Journal of Econometrics, 146(1), 170-184.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
