% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq.test.lvsl_1.R
\name{dq.test.lvsl_1}
\alias{dq.test.lvsl_1}
\title{Sequential Test for breaks over a range of quantiles}
\usage{
dq.test.lvsl_1(y, x, q.L, q.R, n.size = 1, vec.date)
}
\arguments{
\item{y}{A vector of dependent variables (NT x 1)}

\item{x}{A matrix of regressors (NT x p)}

\item{q.L}{The lower end of the range of quantiles}

\item{q.R}{The upper end of the range of quantiles}

\item{n.size}{The size of cross sections (N)}

\item{vec.date}{Estimated break dates under the null hypotheis (Lx1)}
}
\value{
A value of the DQ test statristics
}
\description{
Sequential Test for breaks over a range of quantiles
}
\details{
This procedure tests for the existence of L breaks against L+1 breaks
over an interval of quantiles:
H0: L breaks v.s.H1: L+1 breaks.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
