% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq.test.0vs1.R
\name{sq.test.0vs1}
\alias{sq.test.0vs1}
\title{Break test for a single regression quantile}
\usage{
sq.test.0vs1(y, x, v.tau, n.size = 1)
}
\arguments{
\item{y}{A vector of dependent variables (NT x 1)}

\item{x}{A matrix of regressors (NT x p)}

\item{v.tau}{A probablity at which a conditional quantile is estimated.}

\item{n.size}{The size of cross sections (N)}
}
\value{
The value of the SQ test
}
\description{
To test the existence of a break in a single conditional quantile.
}
\details{
This function tests a single break in a linear quantile regression
}
\references{
Koenker, R., and Bassett Jr, G. (1978).
"Regression quantiles". Econometrica, 46(1), 33-50.

Qu, Z. (2008).
"Testing for structural change in regression quantiles."
Journal of Econometrics, 146(1), 170-184.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
