% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq.test.lvsl_1.R
\name{sq.test.lvsl_1}
\alias{sq.test.lvsl_1}
\title{Break test for a single regression quantile}
\usage{
sq.test.lvsl_1(y, x, v.tau, n.size = 1, vec.date)
}
\arguments{
\item{y}{A vector of dependent variables (NT x 1)}

\item{x}{A matrix of regressors (NT x p)}

\item{v.tau}{A quantile of interest}

\item{n.size}{The size of cross sections ($N$)}

\item{vec.date}{A vector of break dates estimated under the null hypothesis}
}
\value{
TBa
}
\description{
Break test for a single regression quantile
}
\details{
This function tests the null hypothesis of L breaks
againsts the alternative hypothesis of L+1 breaks
in a single conditional quantile.
}
\references{
Oka, T., and Qu, Z. (2011).
"Estimating structural changes in regression quantiles."
Journal of Econometrics, 162(2), 248-267.

Qu, Z. (2008).
"Testing for structural change in regression quantiles."
Journal of Econometrics, 146(1), 170-184.
}
\author{
Tatsushi Oka and Zhongjun Qu
}
